requirements_extended.txt

streamlit
python-docx
beautifulsoup4
requests
